backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.12
fonttools==4.56.0
h11==0.14.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
mysql-connector-python==9.2.0
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pydantic==2.11.1
pydantic_core==2.33.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
